refactor: update types and declare directives folder in the module.ts file
