+++ src/runtime/composables/use-form.ts
@@ -1,15 +1,10 @@
 import type { z } from "zod"
-import { useAbstractForm } from "../../lib/core/composables/use-abstract-form"
 import type { AbstractSchema, UseFormConfiguration } from "../../types/types-api"
 import type { DeepPartial, GenericForm } from "../../types/types-core"
 import type { TypeWithNullableDynamicKeys } from "../../types/types-zod"
 import type { UnwrapZodObject, UseFormConfigurationWithZod } from "../../types/types-zod-adapter"
 import { zodAdapter } from "../adapters/zod"
-
-type UseFormReturnType<Form extends GenericForm, GetValueFormType extends GenericForm = Form> = ReturnType<typeof useAbstractForm<
-  Form,
-  GetValueFormType
->>
+import { useAbstractForm, type UseAbstractFormReturnType } from "./use-abstract-form"
 
 // Overload the useForm type definition to signal that zod schemas have 1st class support
 export function useForm<
@@ -22,7 +17,7 @@ export function useForm<
     AbstractSchema<Form, GetValueFormType>,
     DeepPartial<Form>
   >,
-): UseFormReturnType<Form, GetValueFormType>
+): UseAbstractFormReturnType<Form, GetValueFormType>
 export function useForm<
   Schema extends z.ZodObject<z.ZodRawShape>,
   GetValueFormType extends GenericForm = TypeWithNullableDynamicKeys<Schema>
@@ -31,7 +26,7 @@ export function useForm<
     Schema,
     DeepPartial<z.infer<UnwrapZodObject<Schema>>>
   >,
-): UseFormReturnType<z.infer<UnwrapZodObject<Schema>>, GetValueFormType>
+): UseAbstractFormReturnType<z.infer<UnwrapZodObject<Schema>>, GetValueFormType>
 export function useForm<
   Schema extends z.ZodSchema<unknown>,
   Form extends GenericForm = z.infer<UnwrapZodObject<Schema>>,
@@ -46,7 +41,7 @@ export function useForm<
     Schema,
     DeepPartial<z.infer<UnwrapZodObject<Schema>>>
   >,
-): UseFormReturnType<Form, GetValueFormType> {
+): UseAbstractFormReturnType<Form, GetValueFormType> {
   function isZodType(value: unknown): value is z.ZodType {
     return typeof value === "object" && value !== null && "_def" in value
   }
