+++ src/runtime/plugins/xmodel.ts
@@ -1,5 +1,5 @@
 import { defineNuxtPlugin } from "nuxt/app"
-import { vXModelDynamic } from "../../lib/core/directives/xmodel"
+import { vXModelDynamic } from "../directives/xmodel"
 
 export default defineNuxtPlugin((nuxtApp) => {
   nuxtApp.vueApp.directive("xmodel", vXModelDynamic)
